extract: bundle a flag block's row routing, and split the usage label tests
Merging S-163's alternation rows beside S-155's argparse gate took
emit_flags_block to eight arguments; RowRouting carries the three row facts.
usage.rs reached 804 code lines, so its label and program-name predicates
move to usage_label.rs. Xvfb gains 24 value names, fc-scan drops a group
label that repeated its own description.
